+++ pkg/virt-launcher-common/notify-client/BUILD.bazel
@@ -1,0 +1,25 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["client.go"],
+    importpath = "kubevirt.io/kubevirt/pkg/virt-launcher-common/notify-client",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/apimachinery/wait:go_default_library",
+        "//pkg/ephemeral-disk-utils:go_default_library",
+        "//pkg/handler-launcher-com:go_default_library",
+        "//pkg/handler-launcher-com/notify/info:go_default_library",
+        "//pkg/handler-launcher-com/notify/v1:go_default_library",
+        "//pkg/util/net/grpc:go_default_library",
+        "//pkg/virt-launcher-common/api:go_default_library",
+        "//staging/src/kubevirt.io/api/core/v1:go_default_library",
+        "//staging/src/kubevirt.io/client-go/log:go_default_library",
+        "//vendor/k8s.io/api/core/v1:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/watch:go_default_library",
+        "//vendor/k8s.io/client-go/tools/reference:go_default_library",
+        "@org_golang_google_grpc//:go_default_library",
+    ],
+)
