+++ pkg/virt-launcher/notify-client/BUILD.bazel
@@ -6,13 +6,8 @@ go_library(
     importpath = "kubevirt.io/kubevirt/pkg/virt-launcher/notify-client",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/apimachinery/wait:go_default_library",
-        "//pkg/ephemeral-disk-utils:go_default_library",
-        "//pkg/handler-launcher-com:go_default_library",
-        "//pkg/handler-launcher-com/notify/info:go_default_library",
-        "//pkg/handler-launcher-com/notify/v1:go_default_library",
-        "//pkg/util/net/grpc:go_default_library",
         "//pkg/virt-launcher-common/api:go_default_library",
+        "//pkg/virt-launcher-common/notify-client:go_default_library",
         "//pkg/virt-launcher/metadata:go_default_library",
         "//pkg/virt-launcher/virtwrap/agent-poller:go_default_library",
         "//pkg/virt-launcher/virtwrap/cli:go_default_library",
@@ -21,14 +16,9 @@ go_library(
         "//pkg/virt-launcher/virtwrap/util:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
-        "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/util/json:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/watch:go_default_library",
-        "//vendor/k8s.io/client-go/tools/reference:go_default_library",
         "//vendor/libvirt.org/go/libvirt:go_default_library",
-        "@org_golang_google_grpc//:go_default_library",
     ],
 )
